refactor: structre refinement for further CI
