+++ .gitignore
@@ -49,7 +49,6 @@ coverage.xml
 .pytest_cache/
 
 # Virtual environments
-.env
 .venv
 env/
 venv/
