--- docker/Dockerfile
@@ -1,44 +1,0 @@
-# Use Python 3.11 slim image for better performance
-FROM python:3.11-slim
-
-# Set working directory
-WORKDIR /app
-
-# Set environment variables
-ENV PYTHONDONTWRITEBYTECODE=1 \
-    PYTHONUNBUFFERED=1 \
-    PYTHONPATH=/app
-
-# Install system dependencies
-RUN apt-get update && apt-get install -y \
-    gcc \
-    g++ \
-    && rm -rf /var/lib/apt/lists/*
-
-# Copy requirements first for better caching
-COPY ../final/requirements.txt .
-
-# Install Python dependencies
-RUN pip install --no-cache-dir --upgrade pip && \
-    pip install --no-cache-dir -r requirements.txt
-
-# Copy application files
-COPY ../final/*.py ./
-COPY ../final/.env ./
-COPY ../final/prompts/ ./prompts/
-COPY ../final/static/ ./static/
-COPY ../final/templates/ ./templates/
-COPY ../final/browser_test.html ./
-
-# Create necessary directories
-RUN mkdir -p reports static/css static/js templates
-
-# Expose port
-EXPOSE 8000
-
-# Health check
-HEALTHCHECK --interval=30s --timeout=30s --start-period=5s --retries=3 \
-    CMD curl -f http://localhost:8000/health || exit 1
-
-# Run the application
-CMD ["python", "run_server.py"]
