--- docker/docker-compose.override.yml
@@ -1,20 +1,0 @@
-# Development overrides for docker-compose
-version: '3.8'
-
-services:
-  land-analysis-api:
-    build:
-      context: .
-      dockerfile: Dockerfile  # Use development Dockerfile
-    volumes:
-      # Mount source code for development (hot reload)
-      - ../final:/app:ro
-      - ./reports:/app/reports
-    environment:
-      # Development environment variables
-      - DEBUG=true
-      - LOG_LEVEL=debug
-    ports:
-      - "8000:8000"
-      - "8001:8001"  # Additional port for debugging
-    command: ["python", "run_server.py"]
