--- docker/docker-compose.yml
@@ -1,30 +1,0 @@
-version: '3.8'
-
-services:
-  land-analysis-api:
-    build:
-      context: .
-      dockerfile: Dockerfile
-    container_name: land-analysis-msa
-    ports:
-      - "8000:8000"
-    environment:
-      - KNOWLEDGE_BASE_ID=1QN1H9SBJM
-      - AWS_REGION=ap-northeast-2
-      - MODEL_ID=anthropic.claude-3-haiku-20240307-v1:0
-      - BIZINFO_API_KEY=VtdTy4
-    volumes:
-      - ./reports:/app/reports
-    restart: unless-stopped
-    healthcheck:
-      test: ["CMD", "curl", "-f", "http://localhost:8000/health"]
-      interval: 30s
-      timeout: 10s
-      retries: 3
-      start_period: 40s
-    networks:
-      - land-analysis-network
-
-networks:
-  land-analysis-network:
-    driver: bridge
