--- final/.gitignore
@@ -1,5 +1,0 @@
-.env
-__pycache__
-reports
-static
-past_codes
