--- final/templates/error.html
@@ -1,22 +1,0 @@
-<!DOCTYPE html>
-<html lang="ko">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>오류 - 토지 분석 AI 서비스</title>
-    <link href="/static/css/style.css" rel="stylesheet">
-</head>
-<body>
-    <div class="container">
-        <div class="error-page">
-            <div class="error-icon">❌</div>
-            <h1>오류가 발생했습니다</h1>
-            <p class="error-message">{{ error }}</p>
-            
-            <div class="error-actions">
-                <a href="/" class="btn btn-primary">홈으로 돌아가기</a>
-            </div>
-        </div>
-    </div>
-</body>
-</html>
