--- final/templates/index.html
@@ -1,68 +1,0 @@
-<!DOCTYPE html>
-<html lang="ko">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>토지 분석 AI 서비스</title>
-    <link href="/static/css/style.css" rel="stylesheet">
-</head>
-<body>
-    <div class="container">
-        <header class="header">
-            <h1>🏡 토지 분석 AI 서비스</h1>
-            <p>AI 기반 토지 분석 및 정책 정보를 제공합니다</p>
-        </header>
-
-        <main class="main-content">
-            <div class="form-container">
-                <h2>토지 정보 입력</h2>
-                <form action="/analyze" method="post" class="analysis-form">
-                    <div class="form-group">
-                        <label for="land_data">토지 데이터:</label>
-                        <textarea 
-                            id="land_data" 
-                            name="land_data" 
-                            rows="6" 
-                            placeholder="예: '주소': '대구광역시 중구 동인동1가 2-1', '지목': '대', '용도지역': '중심상업지역', '용도지구': '지정되지않음', '토지이용상황': '업무용', '지형고저': '평지', '형상': '세로장방', '도로접면': '광대소각', '공시지가': 3735000"
-                            required
-                        ></textarea>
-                    </div>
-                    
-                    <button type="submit" class="submit-btn">
-                        🔍 분석 시작
-                    </button>
-                </form>
-            </div>
-
-            <div class="info-section">
-                <h3>📋 입력 형식 안내</h3>
-                <div class="info-card">
-                    <p><strong>필수 정보:</strong></p>
-                    <ul>
-                        <li>주소: 토지의 정확한 주소</li>
-                        <li>지목: 토지의 지목 분류</li>
-                        <li>용도지역: 도시계획상 용도지역</li>
-                        <li>공시지가: 해당 연도 공시지가</li>
-                    </ul>
-                </div>
-                
-                <div class="info-card">
-                    <p><strong>분석 내용:</strong></p>
-                    <ul>
-                        <li>🤖 AI 기반 토지 특성 분석</li>
-                        <li>🏛️ 관련 정부 지원 정책 검색</li>
-                        <li>📊 투자 가치 및 개발 가능성 평가</li>
-                        <li>📄 전문적인 분석 보고서 생성</li>
-                    </ul>
-                </div>
-            </div>
-        </main>
-
-        <footer class="footer">
-            <p>&copy; 2025 토지 분석 AI 서비스. AI 기반 분석 결과이므로 참고용으로만 사용하세요.</p>
-        </footer>
-    </div>
-
-    <script src="/static/js/main.js"></script>
-</body>
-</html>
