+++ src/.env
@@ -1,0 +1,4 @@
+KNOWLEDGE_BASE_ID=1QN1H9SBJM
+AWS_REGION=ap-northeast-2
+MODEL_ID=anthropic.claude-3-haiku-20240307-v1:0
+BIZINFO_API_KEY=VtdTy4
