+++ src/requirements.txt
@@ -1,0 +1,83 @@
+annotated-types==0.7.0
+anyio==4.10.0
+asyncio==4.0.0
+attrs==25.3.0
+aws-requests-auth==0.4.3
+beautifulsoup4==4.13.4
+boto3==1.40.6
+botocore==1.40.6
+certifi==2025.8.3
+charset-normalizer==3.4.3
+click==8.2.1
+datetime==5.5
+dill==0.4.0
+docstring-parser==0.17.0
+dotenv==0.9.9
+fastapi==0.116.1
+h11==0.16.0
+html5lib==1.1
+httpcore==1.0.9
+httpx==0.28.1
+httpx-sse==0.4.1
+idna==3.10
+importlib-metadata==8.7.0
+jinja2==3.1.6
+jmespath==1.0.1
+jsonschema==4.25.0
+jsonschema-specifications==2025.4.1
+lxml==6.0.0
+markdown-it-py==3.0.0
+markdownify==1.2.0
+markupsafe==3.0.2
+mcp==1.12.4
+mdurl==0.1.2
+mpmath==1.3.0
+opentelemetry-api==1.36.0
+opentelemetry-instrumentation==0.57b0
+opentelemetry-instrumentation-threading==0.57b0
+opentelemetry-sdk==1.36.0
+opentelemetry-semantic-conventions==0.57b0
+packaging==25.0
+pillow==11.3.0
+prompt-toolkit==3.0.51
+pydantic==2.11.7
+pydantic-core==2.33.2
+pydantic-settings==2.10.1
+pygments==2.19.2
+pyjwt==2.10.1
+python-dateutil==2.9.0.post0
+python-dotenv==1.1.1
+python-json-logger==2.0.7
+python-multipart==0.0.20
+pytz==2025.2
+readabilipy==0.3.0
+referencing==0.36.2
+regex==2025.7.34
+requests==2.32.4
+rich==14.1.0
+rpds-py==0.27.0
+s3transfer==0.13.1
+setuptools==80.9.0
+six==1.17.0
+slack-bolt==1.23.0
+slack-sdk==3.36.0
+sniffio==1.3.1
+soupsieve==2.7
+sse-starlette==3.0.2
+starlette==0.47.2
+strands-agents==1.4.0
+strands-agents-tools==0.2.3
+sympy==1.14.0
+tenacity==9.1.2
+typing
+typing-extensions
+typing-inspection
+urllib3==2.5.0
+uvicorn==0.35.0
+watchdog==6.0.0
+wcwidth==0.2.13
+webencodings==0.5.1
+weasyprint
+wrapt==1.17.2
+zipp==3.23.0
+zope-interface==7.2
